+++ src/euphorie/upgrade/deployment/v1/20230405090135_register_redirection_storage/upgrade.py
@@ -1,0 +1,8 @@
+from ftw.upgrade import UpgradeStep
+
+
+class RegisterRedirectionStorage(UpgradeStep):
+    """Register redirection storage."""
+
+    def __call__(self):
+        self.install_upgrade_profile()